% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickglobe.R
\name{qg_style}
\alias{qg_style}
\title{Style a quickglobe}
\usage{
qg_style(
  quickglobe,
  palette = NULL,
  titleFontFamily = NULL,
  formatLegendTicks = NULL,
  legendCells = NULL
)
}
\arguments{
\item{quickglobe}{An object of class 'quickglobe'}

\item{palette}{A D3 color palette}

\item{titleFontFamily}{font family of title label}

\item{formatLegendTicks}{D3 formatter to legend ticks}

\item{legendCells}{number of legend cells}
}
\value{
quickglobe
}
\description{
Style a quickglobe
}
