% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickglobe.R
\name{quickglobe}
\alias{quickglobe}
\title{Spin the Globe!}
\usage{
quickglobe(.data, isoIdCol, valueCol, colorRamp1 = "#fff7ec",
  colorRamp2 = "#7f0000")
}
\arguments{
\item{.data}{A data frame with a at least two columns.}

\item{isoIdCol}{Column name of an ISO-3166 formatted country code.}

\item{valueCol}{Column name of the value to shade.}

\item{colorRamp1}{The bottom of the rendered color ramp.}

\item{colorRamp2}{The top of the rendered color ramp.}
}
\description{
Render an interactive 3D globe.
}
