% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickglobe.R
\name{quickglobe}
\alias{quickglobe}
\title{Spin the Globe!}
\usage{
quickglobe(data, identifier, fill, legend = TRUE, title = NULL)
}
\arguments{
\item{data}{A data frame with a at least two columns.}

\item{identifier}{Column name of country identifier.}

\item{fill}{Column name of the value to shade.}

\item{legend}{Whether to include a legend}

\item{title}{Viz title, centered on top of globe.}
}
\description{
Render an interactive 3D globe.
}
